feat(ui): extract form/search etc pure utils (#28704)
* feat(ui): extract form/search/query/advanced-search/explore/classification/tags/glossary/knowledge-page pure utils

Create *PureUtils.ts files for framework-independent logic extracted from:
- formUtils.tsx → formPureUtils.ts (transformErrors, getPopupContainer, etc.)
- SearchUtils.tsx → SearchPureUtils.ts (getEntityTypeFromSearchIndex, parseBucketsData, getTermQuery)
- QueryBuilderUtils.tsx → QueryBuilderPureUtils.ts (all non-JSX query builder logic)
- AdvancedSearchUtils.tsx → AdvancedSearchPureUtils.ts (filter/option builders)
- ExploreUtils.tsx → ExplorePureUtils.ts (URL parsing, tree/filter utils)
- ClassificationUtils.tsx → ClassificationPureUtils.ts (getDeleteButtonData, getClassificationInfo)
- TagsUtils.tsx → TagsPureUtils.ts (tag creation/query/redirect helpers)
- GlossaryUtils.tsx → GlossaryPureUtils.ts (tree builders, breadcrumbs, validators)
- KnowledgePageUtils.tsx → KnowledgePagePureUtils.ts (page hierarchy/path/tree utils)

Original files retain JSX/singleton functions and re-export all pure symbols for backward compat.


* fix(ui): import getTermQuery from SearchPureUtils and getQueryFilterToIncludeApprovedTerm from GlossaryPureUtils

TagsPureUtils was importing from SearchUtils.tsx (React/SVG-heavy) and TagsUtils.tsx
from GlossaryUtils instead of their pure counterparts, defeating the purity goal.


---------
