+++ openmetadata-ui/src/main/resources/ui/src/utils/SearchUtils.tsx
@@ -13,7 +13,6 @@
 
 import { SearchOutlined } from '@ant-design/icons';
 import { Button, Typography } from 'antd';
-import { Bucket } from 'Models';
 import { Link } from 'react-router-dom';
 import { ReactComponent as GlossaryTermIcon } from '../assets/svg/book.svg';
 import { ReactComponent as IconChart } from '../assets/svg/chart.svg';
@@ -35,14 +34,19 @@ import {
   Option,
   SearchSuggestions,
 } from '../context/GlobalSearchProvider/GlobalSearchSuggestions/GlobalSearchSuggestions.interface';
-import { EntityType, FqnPart } from '../enums/entity.enum';
+import { FqnPart } from '../enums/entity.enum';
 import { SearchIndex } from '../enums/search.enum';
 import { SearchSourceAlias } from '../interface/search.interface';
 import { getPartialNameFromTableFQN } from './FqnUtils';
 import i18n from './i18next/LocalUtil';
-import { ElasticsearchQuery } from './QueryBuilderUtils';
 import searchClassBase from './SearchClassBase';
 import serviceUtilClassBase from './ServiceUtilClassBase';
+
+export {
+  getEntityTypeFromSearchIndex,
+  getTermQuery,
+  parseBucketsData,
+} from './SearchPureUtils';
 
 export const getGroupLabel = (index: string) => {
   let label = '';
@@ -249,243 +253,4 @@ export const filterOptionsByIndex = (
     .filter((option) => option._source?.entityType === entityType)
     .map((option) => option._source)
     .slice(0, maxItemsPerType);
-};
-
-export const getEntityTypeFromSearchIndex = (searchIndex: string) => {
-  const commonAssets: Record<string, EntityType> = {
-    [SearchIndex.TABLE]: EntityType.TABLE,
-    [SearchIndex.PIPELINE]: EntityType.PIPELINE,
-    [SearchIndex.DASHBOARD]: EntityType.DASHBOARD,
-    [SearchIndex.MLMODEL]: EntityType.MLMODEL,
-    [SearchIndex.TOPIC]: EntityType.TOPIC,
-    [SearchIndex.CONTAINER]: EntityType.CONTAINER,
-    [SearchIndex.STORED_PROCEDURE]: EntityType.STORED_PROCEDURE,
-    [SearchIndex.DASHBOARD_DATA_MODEL]: EntityType.DASHBOARD_DATA_MODEL,
-    [SearchIndex.SEARCH_INDEX]: EntityType.SEARCH_INDEX,
-    [SearchIndex.DATABASE_SCHEMA]: EntityType.DATABASE_SCHEMA,
-    [SearchIndex.DATABASE_SERVICE]: EntityType.DATABASE_SERVICE,
-    [SearchIndex.MESSAGING_SERVICE]: EntityType.MESSAGING_SERVICE,
-    [SearchIndex.DASHBOARD_SERVICE]: EntityType.DASHBOARD_SERVICE,
-    [SearchIndex.PIPELINE_SERVICE]: EntityType.PIPELINE_SERVICE,
-    [SearchIndex.ML_MODEL_SERVICE]: EntityType.MLMODEL_SERVICE,
-    [SearchIndex.STORAGE_SERVICE]: EntityType.STORAGE_SERVICE,
-    [SearchIndex.SEARCH_SERVICE]: EntityType.SEARCH_SERVICE,
-    [SearchIndex.GLOSSARY_TERM]: EntityType.GLOSSARY_TERM,
-    [SearchIndex.TAG]: EntityType.TAG,
-    [SearchIndex.DATABASE]: EntityType.DATABASE,
-    [SearchIndex.DOMAIN]: EntityType.DOMAIN,
-    [SearchIndex.DATA_PRODUCT]: EntityType.DATA_PRODUCT,
-    [SearchIndex.API_COLLECTION]: EntityType.API_COLLECTION,
-    [SearchIndex.API_ENDPOINT]: EntityType.API_ENDPOINT,
-    [SearchIndex.METRIC]: EntityType.METRIC,
-    [SearchIndex.API_SERVICE]: EntityType.API_SERVICE,
-  };
-
-  return commonAssets[searchIndex] || null; // Return null if not found
-};
-
-/**
- * Parse bucket data from aggregation responses into a format suitable for select fields
- * @param buckets - The bucket data from aggregation response
- * @param sourceFields - Optional string representing dot-notation path to extract values
- * @returns An array of objects with value and title properties
- */
-export const parseBucketsData = (
-  buckets: Array<Bucket>,
-  sourceFields?: string,
-  sourceFieldOptionType?: {
-    label: string;
-    value: string;
-  }
-) => {
-  if (sourceFieldOptionType) {
-    return buckets.map((bucket) => {
-      const topHitsData = (bucket as Record<string, unknown>)[
-        'top_hits#top'
-      ] as
-        | {
-            hits?: {
-              hits?: Array<{
-                _source?: Record<string, unknown>;
-              }>;
-            };
-          }
-        | undefined;
-      const data = topHitsData?.hits?.hits?.[0]?._source;
-
-      return {
-        title: data?.[sourceFieldOptionType.label] as string,
-        value: data?.[sourceFieldOptionType.value] as string,
-      };
-    });
-  }
-
-  return buckets.map((bucket) => {
-    const topHitsSource = (
-      (bucket as Record<string, unknown>)['top_hits#top'] as
-        | {
-            hits?: {
-              hits?: Array<{
-                _source?: Record<string, unknown>;
-              }>;
-            };
-          }
-        | undefined
-    )?.hits?.hits?.[0]?._source;
-
-    const actualValue =
-      sourceFields && topHitsSource
-        ? sourceFields
-            .split('.')
-            .reduce(
-              (obj: unknown, key: string): unknown =>
-                obj && typeof obj === 'object' && obj !== null && key in obj
-                  ? (obj as Record<string, unknown>)[key]
-                  : undefined,
-              topHitsSource
-            ) ?? bucket.key
-        : bucket.key;
-
-    return {
-      value: actualValue,
-      title: bucket.label ?? actualValue,
-    };
-  });
-};
-
-/**
- * Generic term query builder from object
- * Creates an Elasticsearch query filter structure from field-value pairs
- * @param terms - Record of field names and their values, or mixed query configuration
- * @param queryType - Type of boolean query: 'must' | 'must_not' | 'should' | 'should_not'
- * @param minimumShouldMatch - Minimum number of should clauses that must match (only for 'should')
- * @param wildcardTerms - Optional record for wildcard queries
- * @returns Query filter object for searchQuery API
- */
-const NESTED_FIELDS = ['owners'];
-
-const getNestedPath = (field: string): string | undefined => {
-  return NESTED_FIELDS.find((nested) => field.startsWith(`${nested}.`));
-};
-
-const wrapTermQuery = (
-  field: string,
-  value: string | number | boolean,
-  nestedPath?: string
-): ElasticsearchQuery => {
-  const termQuery: ElasticsearchQuery = { term: { [field]: value } };
-  if (nestedPath) {
-    return { nested: { path: nestedPath, query: termQuery } };
-  }
-
-  return termQuery;
-};
-
-export const getTermQuery = (
-  terms: Record<string, string | string[] | number | boolean>,
-  queryType: 'must' | 'must_not' | 'should' | 'should_not' = 'must',
-  minimumShouldMatch?: number,
-  options?: {
-    wildcardTerms?: Record<string, string>;
-    wildcardShouldQueries?: Record<string, string>;
-    mustNotTerms?: Record<string, string | string[] | number | boolean>;
-    matchTerms?: Record<string, string | number | boolean>;
-    wildcardMustNotQueries?: Record<string, string | string[]>;
-  }
-) => {
-  const termQueries = Object.entries(terms).flatMap(([field, value]) => {
-    const nestedPath = getNestedPath(field);
-    if (Array.isArray(value)) {
-      return value.map((v) => wrapTermQuery(field, v, nestedPath));
-    }
-
-    return wrapTermQuery(field, value, nestedPath);
-  });
-
-  const wildcardQueries = options?.wildcardTerms
-    ? Object.entries(options.wildcardTerms).map(([field, value]) => ({
-        wildcard: { [field]: value },
-      }))
-    : [];
-
-  const mustNotQueries = options?.mustNotTerms
-    ? Object.entries(options.mustNotTerms).flatMap(([field, value]) => {
-        const nestedPath = getNestedPath(field);
-        if (Array.isArray(value)) {
-          return value.map((v) => wrapTermQuery(field, v, nestedPath));
-        }
-
-        return wrapTermQuery(field, value, nestedPath);
-      })
-    : [];
-
-  const matchQueries = options?.matchTerms
-    ? Object.entries(options.matchTerms).map(([field, value]) => ({
-        match: { [field]: value },
-      }))
-    : [];
-
-  const allQueries: ElasticsearchQuery[] = [
-    ...termQueries,
-    ...wildcardQueries,
-    ...matchQueries,
-  ];
-
-  // Handle wildcardShouldQueries - creates a nested bool with should clauses
-  if (
-    options?.wildcardShouldQueries &&
-    Object.keys(options.wildcardShouldQueries).length > 0
-  ) {
-    const shouldWildcardQueries = Object.entries(
-      options.wildcardShouldQueries
-    ).map(([field, value]) => ({
-      wildcard: { [field]: value },
-    }));
-
-    allQueries.push({
-      bool: {
-        should: shouldWildcardQueries,
-        minimum_should_match: 1,
-      },
-    });
-  }
-
-  // Define type for Elasticsearch bool query structure
-  type ESBoolQuery = Record<string, ElasticsearchQuery[] | number> & {
-    must_not?: ElasticsearchQuery[];
-    minimum_should_match?: number;
-  };
-
-  const boolQuery: ESBoolQuery = {
-    [queryType]: allQueries,
-  };
-
-  // Handle wildcardMustNotQueries
-  const wildcardMustNotQueries = options?.wildcardMustNotQueries
-    ? Object.entries(options.wildcardMustNotQueries).flatMap(
-        ([field, value]) => {
-          if (Array.isArray(value)) {
-            return value.map((v) => ({ wildcard: { [field]: v } }));
-          }
-
-          return { wildcard: { [field]: value } };
-        }
-      )
-    : [];
-
-  const allMustNotQueries = [...mustNotQueries, ...wildcardMustNotQueries];
-
-  if (allMustNotQueries.length > 0) {
-    boolQuery.must_not = allMustNotQueries;
-  }
-
-  if (queryType === 'should' && minimumShouldMatch !== undefined) {
-    boolQuery.minimum_should_match = minimumShouldMatch;
-  }
-
-  return {
-    query: {
-      bool: boolQuery,
-    },
-  };
 };
